Extract world diagnostics helpers
